Suppress compiler nag.
git-svn-id: https://simulationcraft.googlecode.com/svn/trunk@15464 3b4652a1-8050-0410-ac47-3d40261b0f8b

+++ engine/report/sc_chart.cpp
@@ -1460,7 +1460,7 @@ std::string chart::reforge_dps( player_t* p )
   if ( num_stats == 2 )
   {
     int num_points = ( int ) pd.size();
-    int range = pd[ num_points - 1 ][ 0 ].value;
+    int range = static_cast< int >( pd[ num_points - 1 ][ 0 ].value );
     std::vector<stat_e> stat_indices = p -> sim -> reforge_plot -> reforge_plot_stat_indices;
     plot_data_t& baseline = pd[ num_points / 2 ][ 2 ];
     double min_delta = baseline.value - ( min_dps - baseline.error / 2 );
